pyproject.toml
[project]
name = "org-demo"
version = "0.1.0"
description = "Create executable demo documents using org-mode as a backend"
authors = [
    {name = "Juan Gonzalez",email = "juang@wolfram.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "python-emacs",
]

[project.scripts]
org-demo = "org_demo.cli:cli"

[tool.poetry]
packages = [{include = "org_demo", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
